+++ Changelog.md
@@ -7,3 +7,10 @@ Pre-release January 18 2021
   - seen in main.py
   - subject to future changes
 - edited README
+
+Pre-release january 27 2021
+- Created interactive weapons and food inventories
+- Moved inventory v1 to original_inventory.py
+- Made locations dictionary
+- Made characters dictionary
+- Made all dictionaries into an interactive directory system